requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chart-studio==1.0.0
Click==7.0
colorama==0.4.3
dash==0.29.0
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.2.3
dash-table==4.4.1
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
neat-python==0.92
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pickleshare==0.7.5
plotly==4.5.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
ta==0.5.17
tensorboard>=2.1.0
tensorflow-estimator>=2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==2.2.0
